Add Babel transformation step to JS file compilation
This adds a Babel transformation step when compiling JS
files. It uses the @babel/preset-env plugin to enable a
baseline set of ES6 features. These changes also make
use of Promises instead of callback functions so that the
`compile` and `compile-js-dev` tasks properly fail if one
of the input files fails to compile.